Restructure node module in ungraph and digraph, chaneg ungraph to use std's BinaryHeap
