+++ src/digraph/node/algo/bfs.rs
