+++ src/digraph/node/algo/method.rs
