+++ src/digraph/node/algo/mod.rs
@@ -1,0 +1,9 @@
+use super::*;
+
+pub mod bfs;
+pub mod dfs;
+pub mod pfs;
+pub mod order;
+
+mod path;
+mod method;
