+++ src/digraph/node/algo/order.rs
