+++ src/digraph/node/algo/path.rs
