+++ src/digraph/node/algo/pfs.rs
