+++ src/ungraph/node/adjacent.rs
@@ -1,5 +1,32 @@
 use super::*;
 
+#[derive(Clone)]
+pub struct WeakNode<K = usize, N = (), E = ()>
+where
+	K: Clone + Hash + Display + PartialEq + Eq,
+	N: Clone,
+	E: Clone,
+{
+	inner: Weak<(K, N, RefCell<Adjacent<K, N, E>>)>,
+}
+
+impl<K, N, E> WeakNode<K, N, E>
+where
+	K: Clone + Hash + Display + PartialEq + Eq,
+	N: Clone,
+	E: Clone,
+{
+	pub fn upgrade(&self) -> Option<Node<K, N, E>> {
+		self.inner.upgrade().map(|inner| Node { inner })
+	}
+
+	pub fn downgrade(node: &Node<K, N, E>) -> Self {
+		WeakNode {
+			inner: Rc::downgrade(&node.inner)
+		}
+	}
+}
+
 pub struct Adjacent<K, N, E>
 where
     K: Clone + Hash + PartialEq + Eq + Display,
