+++ src/ungraph/node/algo/method.rs
